Add files via upload
The Random Forest algorithm is a versatile and widely-used machine learning method for classification and regression tasks. It operates by constructing a multitude of decision trees during training and outputting the class that is the mode of the classes (classification) or mean prediction (regression) of the individual trees. This ensemble approach helps improve accuracy and control over-fitting. Random Forests are robust to noise and handle large datasets efficiently. They can manage thousands of input variables without variable deletion and provide estimates of what variables are important in the classification. Additionally, they have an effective method for estimating missing data and maintain accuracy for unbalanced datasets. This algorithm is beneficial for various applications, including fraud detection, medical diagnosis, and recommendation systems.